% 2005-09-23, Alberto Viglione
%
\name{exampleRFA}
\alias{exampleRFA}
\alias{exampleRFA01}
\alias{exampleRFA02}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Example of package functionality}
\description{
 Examples of Regional Frequency Analysis.
}
\usage{
 exampleRFA01 ()
 exampleRFA02 ()
}
%- maybe also `usage' for other objects documented here.
%\arguments{\item{}{}}
}
%\details{}
\value{\code{exampleRFA01} and \code{exampleRFA02} are about the Regional Frequency Analysis of annual flows in Piemonte and Valle d'Aosta.
\code{exampleRFA01} treats the regionalization of the index-flow.
\code{exampleRFA02} treats the regionalization of the growth-curve.
}
\references{
Viglione A., Claps P., Laio F. (2006) Utilizzo di criteri di prossimit\`a nell'analisi regionale del deflusso annuo, XXX Convegno di Idraulica e Costruzioni Idrauliche - IDRA 2006, Roma, 10-15 Settembre 2006.

Viglione A., Claps P., Laio F. (2006) Water resources assessment and management under water scarcity scenarios, chapter Mean annual runoff estimation in North-Western Italy. CDSU, Milan.

Viglione A. (2007) Metodi statistici non-supervised per la stima di grandezze idrologiche in siti non strumentati, PhD thesis
, In press.
}

\author{Alberto Viglione, e-mail: \email{alviglio@tiscali.it}.}

%\note{.}

%\seealso{}

\examples{
exampleRFA01()
#exampleRFA02()          # it takes some time
}
\keyword{manip}
