% 2005-09-23, Alberto Viglione
%
\name{functionsLaio}
\docType{data}
\alias{df.kgev}
\alias{df.polig}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Data-sample}
\description{
Functions for inversion calculation.
}
\usage{
 data(functionsLaio)
}
\format{Data.frames:

 \code{df.kgev} is a data.frame with the skewness coefficient (first column) and the corresponding shape parameter of the GEV (second column) 
 
 \code{df.polig} represents the poligamma function.
}
%\source{}
%\references{}
%\author{Alberto Viglione}
\note{For information on the package and the Author, and for all the references, see \code{\link{nsRFA}}.}


%\examples{}
\keyword{datasets}
