\name{nsRFA-package}
\alias{nsRFA-package}
\alias{nsRFA}
\docType{package}
\title{
Non-supervised Regional Frequency Analysis
}
\description{
A collection of statistical tools for objective (non-supervised) applications
of the Regional Frequency Analysis methods in hydrology.
}
\details{
\tabular{ll}{
Package: \tab nsRFA\cr
Type: \tab Package\cr
Version: \tab 0.5-2\cr
Date: \tab 2008-04-22\cr
License: \tab GPL (>= 2)\cr
}
The package provides several tools for Regional Frequency Analysis of hydrological variables.
The first version dates to 2006 and was developed in Turin at the Politecnico by Alberto Viglione.

~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Alberto Viglione

Maintainer: Alberto Viglione <alviglio@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{package}
\keyword{htest}
\keyword{distribution}
\keyword{univar}
\keyword{hplot}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
demo(rfa_annualflows01)

demo(rfa_annualflows02)

demo(feh_floods01)

demo(feh_floods02)

demo(feh_floods03)
}
