% 2005-09-23, Alberto Viglione
%
\name{rfa-internal}
\alias{rfa-internal}
\alias{ksampleA2}
\alias{nonparboot}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Internal functions}
\description{
 User-level objects which are for \sQuote{internal} use only
 ksampleA2 provides the Anderson-Darling test statistic;
 nonparboot provides a non-parametric bootstrap.
}
\usage{
 ksampleA2 (x,cod)
 nonparboot (z,n=length(z))
}
%- maybe also `usage' for other objects documented here.
%\arguments{}
\details{
 \code{ksampleA2} provides the Anderson-Darling test statistic;
 \code{nonparboot} provides a non-parametric bootstrap.
}
%\value{}

\author{Alberto Viglione, e-mail: \email{alviglio@tiscali.it}.}

%\note{.}

%\seealso{\code{\link{factor}}, \code{\link{HOMTESTS}}, \code{\link{CLUSTERS}}}

%\examples{}
\keyword{internal}
