% 2005-09-23, Alberto Viglione
%
\name{functionsLaio}
\docType{data}
\alias{kgev}
\alias{polig}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Data-sample}
\description{
Functions for inversion calculation.
}
\usage{
 data(functionsLaio)
}
\format{Data.frames:

 \code{kgev} is a matrix with the skewness coefficient (first column) and the corresponding shape parameter of the GEV (second column) 
 
 \code{polig} represents the poligamma function.
}
%\source{}
\references{

}

\examples{

}
\keyword{datasets}
